refactor: remove util.pool
